test: decouple ignore-fn test from esm.sh; skipbrowser + SKIP_ESMSH for esm.sh tests
- Rewrite ignore-fn test using jspm.io @react-three/fiber (mirrors ignore.test.js) so it no longer depends on esm.sh.
- Mark existing esm.sh tests skipbrowser (CORS) and add SKIP_ESMSH env guard for Node so they can be toggled when esm.sh is unreliable.
