--- generator/test/api/local/esmsh-import/entry.js
@@ -1,3 +1,0 @@
-import lodash from 'https://esm.sh/lodash@4.17.21';
-
-export default lodash;
