+++ generator/test/html/esmsh.skipbrowser.test.js
@@ -2,6 +2,11 @@ import { Generator } from '@jspm/generator';
 import assert from 'assert';
 import { SemverRange } from 'sver';
 
+if (process.env.SKIP_ESMSH) {
+  console.log('Skipping esm.sh test — SKIP_ESMSH set');
+  process.exit(0);
+}
+
 const generator = new Generator({
   mapUrl: new URL('./local/page.html', import.meta.url),
   env: ['production', 'browser'],
