+++ generator/test/providers/esmsh.skipbrowser.test.js
@@ -1,6 +1,11 @@
 import { Generator } from '@jspm/generator';
 import assert from 'assert';
 
+if (process.env.SKIP_ESMSH) {
+  console.log('Skipping esm.sh test — SKIP_ESMSH set');
+  process.exit(0);
+}
+
 const generator = new Generator({
   mapUrl: import.meta.url,
   defaultProvider: 'esm.sh',
